* Adicionada lógica de Despesas e Receitas por usuario. * Ajustados todos repositórios para verificar por usuario. * Adicionada lógica para trazer email a partir do token e buscar usuário logado * Adicionada expcetion para email já cadastrado

+++ src/main/kotlin/br/com/alura/financeiro/service/DespesaService.kt
@@ -6,6 +6,7 @@ import br.com.alura.financeiro.exception.InvalidEnumeratorException
 import br.com.alura.financeiro.exception.NotFoundException
 import br.com.alura.financeiro.model.Categoria
 import br.com.alura.financeiro.model.Despesa
+import br.com.alura.financeiro.model.Usuario
 import br.com.alura.financeiro.repository.DespesaRepository
 import org.springframework.stereotype.Service
 import java.time.LocalDate
@@ -15,25 +16,29 @@ class DespesaService(
     private val repository: DespesaRepository,
     val notFoundMessage: String = "Despesa não encontrada!"
 ) {
-    fun listar(): List<DespesaView> {
-        val despesas = repository.findAll()
+    fun listar(usuario: Usuario): List<DespesaView> {
+        val despesas = repository.findByUsuario_Id(usuario.id!!)
 
         return DespesaView.converter(despesas)
     }
 
-    fun buscarPorId(id: Long): DespesaView{
-        val despesa = repository.findById(id).orElseThrow {NotFoundException(notFoundMessage)}
+    fun buscarPorId(id: Long, usuario: Usuario): DespesaView{
+        val despesa = repository.findByIdAndUsuario_Id(id, usuario.id!!)
 
-        return DespesaView(
+        if(despesa != null) {
+            return DespesaView(
                 id = despesa.id,
                 descricao = despesa.descricao,
                 valor = despesa.valor,
                 data = despesa.data,
                 categoria = despesa.categoria
-        )
+            )
+        } else{
+            throw NotFoundException(notFoundMessage)
+        }
     }
 
-    fun cadastrar(nova: DespesaForm) {
+    fun cadastrar(nova: DespesaForm, usuario: Usuario) {
         if (nova.categoria in Categoria.values()) {
             try {
                 repository.save(
@@ -41,7 +46,8 @@ class DespesaService(
                         descricao = nova.descricao,
                         valor = nova.valor,
                         data = nova.data,
-                        categoria = nova.categoria
+                        categoria = nova.categoria,
+                        usuario = usuario
                     )
                 )
             } catch (e: Exception) {
@@ -52,39 +58,48 @@ class DespesaService(
         }
     }
 
-    fun atualizar(id: Long,  nova: DespesaForm) : DespesaView {
-        val despesa = repository.findById(id).orElseThrow{NotFoundException(notFoundMessage)}
+    fun atualizar(id: Long, nova: DespesaForm, usuario: Usuario) : DespesaView {
+        val despesa = repository.findByIdAndUsuario_Id(id,usuario.id!!)
 
-        despesa.valor = nova.valor
-        despesa.descricao = nova.descricao
-        despesa.data = nova.data
-        despesa.categoria = nova.categoria
+        if(despesa != null) {
+            despesa.valor = nova.valor
+            despesa.descricao = nova.descricao
+            despesa.data = nova.data
+            despesa.categoria = nova.categoria
 
-        return DespesaView(
-            id = despesa.id,
-            descricao = despesa.descricao,
-            valor = despesa.valor,
-            data = despesa.data,
-            categoria = despesa.categoria
-        )
+            return DespesaView(
+                id = despesa.id,
+                descricao = despesa.descricao,
+                valor = despesa.valor,
+                data = despesa.data,
+                categoria = despesa.categoria
+            )
+        } else {
+            throw NotFoundException(notFoundMessage)
+        }
     }
 
-    fun remover(id: Long) {
-        val remove = repository.findById(id).orElseThrow { NotFoundException(notFoundMessage) }
-        repository.delete(remove)
+    fun remover(id: Long, usuario: Usuario) {
+        val remove = repository.findByIdAndUsuario_Id(id, usuario.id!!)
+
+        if(remove != null) {
+            repository.delete(remove)
+        } else {
+            throw NotFoundException(notFoundMessage)
+        }
     }
 
-    fun buscarPorDescricao(descricao: String): List<DespesaView> {
-        val lista = repository.findByDescricaoContaining(descricao)
+    fun buscarPorDescricao(descricao: String, usuario: Usuario): List<DespesaView> {
+        val lista = repository.findByDescricaoContainingAndUsuario_Id(descricao, usuario.id)
 
         return DespesaView.converter(lista)
     }
 
-    fun buscarPorMes(ano: Int, mes: Int): List<DespesaView> {
+    fun buscarPorMes(ano: Int, mes: Int, usuario: Usuario): List<DespesaView> {
         val primeiroDia = LocalDate.of(ano,mes, 1)
         val ultimoDia = primeiroDia.withDayOfMonth(primeiroDia.lengthOfMonth())
 
-        return DespesaView.converter(repository.findByDataBetween(primeiroDia, ultimoDia))
+        return DespesaView.converter(repository.findByDataBetweenAndUsuario_Id(primeiroDia, ultimoDia, usuario.id))
     }
 
 }
