+++ src/main/kotlin/br/com/alura/financeiro/service/ReceitasService.kt
@@ -6,6 +6,7 @@ import br.com.alura.financeiro.exception.InvalidEnumeratorException
 import br.com.alura.financeiro.exception.NotFoundException
 import br.com.alura.financeiro.model.Categoria
 import br.com.alura.financeiro.model.Receita
+import br.com.alura.financeiro.model.Usuario
 import br.com.alura.financeiro.repository.ReceitaRepository
 import org.springframework.stereotype.Service
 import java.time.LocalDate
@@ -15,25 +16,29 @@ class ReceitasService(
     private val repository : ReceitaRepository,
     val notFoundMessage: String = "Receita não encontrada!"
 ) {
-    fun listar(): List<ReceitaView> {
-        val receitas = repository.findAll()
+    fun listar(usuario: Usuario): List<ReceitaView> {
+        val receitas = repository.findByUsuario_Id(usuario.id!!)
 
         return ReceitaView.converter(receitas)
     }
 
-    fun buscaPorId(id: Long): ReceitaView {
-        val receita = repository.findById(id).orElseThrow{NotFoundException(notFoundMessage)}
+    fun buscaPorId(id: Long, usuario: Usuario): ReceitaView {
+        val receita = repository.findByIdAndUsuario_Id(id, usuario.id!!)
 
-        return ReceitaView(
-             id = receita.id,
-             descricao = receita.descricao,
-             valor = receita.valor,
-             data = receita.data,
-             categoria = receita.categoria
-        )
+        if(receita != null) {
+            return ReceitaView(
+                id = receita.id,
+                descricao = receita.descricao,
+                valor = receita.valor,
+                data = receita.data,
+                categoria = receita.categoria
+            )
+        } else{
+            throw NotFoundException(notFoundMessage)
+        }
     }
 
-    fun cadastrar(nova: ReceitaForm) {
+    fun cadastrar(nova: ReceitaForm, usuario: Usuario) {
         if (nova.categoria in Categoria.values()) {
             try {
                 repository.save(
@@ -41,7 +46,8 @@ class ReceitasService(
                         descricao = nova.descricao,
                         valor = nova.valor,
                         data = nova.data,
-                        categoria = nova.categoria
+                        categoria = nova.categoria,
+                        usuario = usuario
                     )
                 )
             } catch (e: Exception) {
@@ -52,43 +58,51 @@ class ReceitasService(
         }
     }
 
-    fun atualizar(id: Long,  nova: ReceitaForm) : ReceitaView {
-        val receita = repository.findById(id).orElseThrow{NotFoundException(notFoundMessage)}
+    fun atualizar(id: Long, nova: ReceitaForm, usuario: Usuario) : ReceitaView {
+        val receita = repository.findByIdAndUsuario_Id(id, usuario.id!!)
 
-        if (nova.categoria in Categoria.values()) {
-            receita.valor = nova.valor
-            receita.descricao = nova.descricao
-            receita.data = nova.data
-            receita.categoria = nova.categoria
-        } else{
-            throw InvalidEnumeratorException("Não é uma categoria válida! "+Categoria.values().toString())
+        if(receita != null) {
+            if (nova.categoria in Categoria.values()) {
+                receita.valor = nova.valor
+                receita.descricao = nova.descricao
+                receita.data = nova.data
+                receita.categoria = nova.categoria
+
+                return ReceitaView(
+                    id = receita.id,
+                    descricao = receita.descricao,
+                    valor = receita.valor,
+                    data = receita.data,
+                    categoria = receita.categoria
+                )
+            } else {
+                throw InvalidEnumeratorException("Não é uma categoria válida! " + Categoria.values().toString())
+            }
+        } else {
+            throw NotFoundException(notFoundMessage)
         }
-
-        return ReceitaView(
-            id = receita.id,
-            descricao = receita.descricao,
-            valor = receita.valor,
-            data = receita.data,
-            categoria = receita.categoria
-        )
     }
 
-    fun remover(id: Long) {
-        val remove = repository.findById(id).orElseThrow { NotFoundException(notFoundMessage) }
-        repository.delete(remove)
+    fun remover(id: Long, usuario: Usuario) {
+        val remove = repository.findByIdAndUsuario_Id(id, usuario.id!!)
+        if (remove != null) {
+            repository.delete(remove)
+        } else {
+            throw NotFoundException(notFoundMessage)
+        }
     }
 
-    fun buscarPorDescricao(descricao: String): List<ReceitaView> {
-        val lista = repository.findByDescricaoContaining(descricao)
+    fun buscarPorDescricao(descricao: String, usuario: Usuario): List<ReceitaView> {
+        val lista = repository.findByDescricaoContainingAndUsuario_Id(descricao, usuario.id!!)
 
         return ReceitaView.converter(lista)
     }
 
-    fun buscarPorMes(ano: Int, mes: Int): List<ReceitaView> {
+    fun buscarPorMes(ano: Int, mes: Int, usuario: Usuario): List<ReceitaView> {
         val primeiroDia = LocalDate.of(ano,mes, 1)
         val ultimoDia = primeiroDia.withDayOfMonth(primeiroDia.lengthOfMonth())
 
-        return ReceitaView.converter(repository.findByDataBetween(primeiroDia, ultimoDia))
+        return ReceitaView.converter(repository.findByDataBetweenAndUsuario_Id(primeiroDia, ultimoDia, usuario.id))
     }
